Trigger climbing works.
Using trigger objects at the top and bottom to climb works.  I would
still much rather figure a way to do it with collision but I can't get
the OnCollisionEnter function to fire.  There is a lot of commented out
code left in the script so that I can go back and test other ways to do
it in the future.