Oculus in but not set up
"V" switches between 1st person, 3rd person and a blank thing that will
be the oculus.  All the oculus stuff is in but it's not playing nice
with the other cameras and is currently doing an all or nothing type
thing.  So I pulled it's camera out to not interfere with other people
working while I figure out how to switch it on and off.  Also moved fist
off main camera and on to the player.